created APIs from No.1 to No.5

+++ src/main/java/com/luckyvicky/hakshikmap/repository/WaitingLogsRepository.java
@@ -3,6 +3,6 @@ import com.luckyvicky.hakshikmap.entity.WaitingLogs;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface WaitingLogsRepository extends JpaRepository<WaitingLogs, Integer> {
+public interface WaitingLogsRepository extends JpaRepository<WaitingLogs, Long> {
 
 }
